pyproject.toml
[project]
name = "voice-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Niklas", email = "niklas@alyst.se" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "boto3>=1.34.0",
    "httpx>=0.25.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
voice-parser = "voice_parser.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
